Fixes #6861. Created new module:unpacking and moved related functions
