+++ tests/unit/test_download.py
@@ -19,10 +19,10 @@ from pip._internal.download import (
     _get_url_scheme,
     parse_content_disposition,
     sanitize_content_filename,
-    unpack_file_url,
     unpack_http_url,
     url_to_path,
 )
+from pip._internal.utils.unpacking import unpack_file_url
 from pip._internal.exceptions import HashMismatch
 from pip._internal.models.link import Link
 from pip._internal.utils.hashes import Hashes
